requirements.txt
aiosmtpd==1.4.2
asgiref==3.5.0
atpublic==2.3
attrs==20.3.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
